added a map of outlets

+++ app/src/main/java/com/euromix/esupervisor/app/model/docEmix/entities/DocEmixDetail.kt
@@ -44,7 +44,7 @@ data class DocEmixDetail(
 data class RowTradeCondition(
     val row: Int,
     val manufacturer: String,
-    val priceIndex: Int,
+    val priceIndex: Float,
     val paymentDeferment: Int,
     val distributionChannel: String,
     val imageURI: String
